pyproject.toml
[project]
name = "usmo"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.1",
    "psutil>=7.1.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.4",
    "rich>=14.0.0",
]

[project.scripts]
usm = "usmo.cli:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
